Extracted rtp receiver payload management to its own class, made video receiver depend on that instead.
Eliminated need for video receiver to talk to its parent. Also we will now determine if the packet is the first one already in the rtp general receiver. The possible downside would be that recovered video packets no longer can be flagged as the first packet, but I don't think that can happen. Even if it can happen, maybe the bit was set anyway at an earlier stage. The tests run fine.

BUG=
TEST=rtp_rtcp_unittests, vie_auto_test, voe_auto_test

Review URL: https://webrtc-codereview.appspot.com/1022011

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@3382 4adac7df-926f-26a2-2b94-8c16560cd09d