Fix fileutils.cc for tests running under Win memory tools.
On Windows, when tests are run under memory tools like
DrMemory and TSan, slashes occur in the path as directory
separators. This makes the fileutils.cc fail to figure out
the project root, making paths to resource files invalid.

By replacing slashes in the path with backslashes for Windows
this can be fixed.

BUG=2318
TEST=running the same command as the bot locally, succeeding after this patch is applied.
R=phoglund@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/2137004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@4640 4adac7df-926f-26a2-2b94-8c16560cd09d