iOS video_capture: start camera in the background.
Camera start is a blocking operation so never a good idea to do on a main
thread, but worse than that is that the guts of WebView appear to be
interacting with capture start in a bad way causing startup to pause for 10s
while a timeout expires.  This change eliminates that 10s delay.

R=noahric@google.com

Review URL: https://webrtc-codereview.appspot.com/10449004

git-svn-id: http://webrtc.googlecode.com/svn/trunk/webrtc@5772 4adac7df-926f-26a2-2b94-8c16560cd09d